test(core): Rewrite busted test in actual Busted test format
